src/groundstation/requirements.txt
absl-py==2.1.0
bleak==0.21.1
certifi==2024.8.30
charset-normalizer==3.3.2
dotmap==1.3.30
idna==3.8
meshtastic==2.4.1
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pexpect==4.9.0
print-color==0.4.6
protobuf==5.28.0
ptyprocess==0.7.0
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-libdispatch==9.2
pyparsing==3.1.4
Pypubsub==4.0.3
PyQRCode==1.2.1
pyserial==3.5
PyYAML==6.0.2
requests==2.32.3
tabulate==0.9.0
urllib3==2.2.2
webencodings==0.5.1
